o Fix setting the datatype for empty connectivity array to a np.int16, initial thought was to optimize, but this led to issues with cross platform handling in integer types, specially windows, hopefully this might fix it. As in previous commit, it was not an issue of large mesh size that was causing this.

+++ uxarray/io/_shapefile.py
@@ -20,7 +20,7 @@ def _read_shpfile(filepath):
     # Initialize as an empty numpy array
     node_lon = np.array([])
     node_lat = np.array([])
-    connectivity = np.empty((0, max_coord_size - 1), dtype=np.int16)
+    connectivity = np.empty((0, max_coord_size - 1))
 
     node_index = 0
